refactor(unity-tests): move async exception helper to dedicated file
